requirements.txt
discord.py==2.3.2
pytube==15.0.0
yt_dlp==2024.11.18
pynacl==1.5.0